pyproject.toml
[tool.poetry]
name = "share-trip-experience"
version = "0.1.0"
description = "app where you can share your journeys with other users"
authors = ["Elizaveta"]
readme = "README.md"
packages = [{include = "share_trip_experience"}]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^3.0.3"
gunicorn = "^22.0.0"
python-dotenv = "^1.0.1"
pymongo = "^4.7.3"
flask-pymongo = "^2.3.0"
flask-login = "^0.6.3"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pytest = "^8.2.2"
coverage = "^7.5.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
